Added ncurses display
Currently, there are 3 windows: a navigation window for a tabbing
feature, a display window to display output of commands and
connections alongside message history, and an input field. Main
has been broken into various files and now main calls
init_display and other functions to present a TUI that further
management of the program will occur in. The display is not very
functional right now and just serves as a blueprint for further
development.