Beginning to reorganize code
Creating global static struct screenState in display.c so that struct
*win doesn't need to be passed to every function. The plan is to make
it so that that data never needs to move outside of the file and
external files can change and print things to the screen via functions
in found in display.h. Also created global static struct netState for
the same purpose in net.c, expect instead of encapsulating the screen
it's encapsulating the networking state and all the connections for the
program.