Fixed multiplexing
Fixed issue in add_to_tab where the text wouldn't be displayed even if
it was being added to the current tab being displayed and read_text
which was trying to read the current connection as opposed to the one
that is passed to it by index. This made it so that it was trying to
read from the tab being displayed instead of the connection that should
be running in the background.