Temporary fix for displaying network msgs
Created variants of add_to_tab and addmsg that take the size of the
message as an argument and use that to define how much gets put into the
tab's text buffer. I opted to create 2 new functions rather than change
the old ones because the old ones are used in a lot of places where it's
just unnecessary to pass in the size.