Implemented generic linked list (Minimally Tested)
The idea is to use this in ScreenState, for storing the list of tabs; in
each tab, for storing the text; and also for storing the network
connections. Only the essential functions have been tested so far, with
the ll->point manipulation functions being left untested. They are
rather simple though, so there shouldn't too many errors.